+++ changelog.md
@@ -4,6 +4,7 @@
 + Added: `sticky-footer`.
 + Added: `parallelogram`.
 + Added: `angled-edges`.
++ Added: `drop-shadow`.
 + Improved: `grid` now works with `@media query` and can make several rows.
 + Improved: rewrite `container` with `css-calc`, now you can combine `container` with other structure tags:
 ```html